requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-huggingface
langchain-groq
langchain-openai
faiss-cpu
pypdf
sentence-transformers